coreboot Makefile: add tag for goes payload to image
Add tag for the goes payload, so the user can tell
what version of goes is built into coreboot(goes-boot).

Signed-off-by: Mark Pleso <mpleso@platinasystems.com>

Makefile.inc: Look for 'go' in sibling directory

Fix the directory path. Since we aren't using system-build, there
are two extra levels of parent to remove. Note that this breaks
building with system-build.

I don't think this is the right way to set the extra version info,
but leave this alone for now.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>